backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.28
alembic>=1.13.1
asyncpg>=0.29.0
pydantic[email]>=2.6.4
python-dotenv>=1.0.1
redis>=5.0.3
structlog>=24.1.0
APScheduler>=3.10.4
pyjwt[crypto]>=2.8.0
cryptography>=42.0.5
slowapi>=0.1.9
firebase-admin>=6.5.0
razorpay>=1.4.1
httpx>=0.27.0
pytz>=2024.1
smartapi-python>=1.4.0
NorenRestApiPy>=0.0.22
fyers-apiv3>=3.1.0

# AI & Analytics
openai>=1.14.0
vaderSentiment>=3.3.2

# Testing & QA
pytest>=8.1.1
pytest-cov>=5.0.0
pytest-asyncio>=0.23.6
pytest-env>=1.1.3
httpx>=0.27.0
bandit>=1.7.8

# Observability
sentry-sdk[fastapi]>=1.43.0